requirements.txt
Flask
dogpile.cache>=0.6
redis